Remove geoImpl, change back to geo
