+++ makefile
@@ -1,7 +1,7 @@
 CFLAGS = -g -Wshadow -std=c++11 -pedantic -Wall -Wextra -Wcast-align -Wcast-qual -Wctor-dtor-privacy -Winit-self -Wlogical-op -Wmissing-declarations -Wmissing-include-dirs -Wnoexcept  -Woverloaded-virtual -Wredundant-decls -Wsign-conversion -Wsign-promo -Wstrict-null-sentinel -Wstrict-overflow=5 -Wswitch-default -Wundef -Wno-unused -Wold-style-cast -Wformat-nonliteral -Wpointer-arith -Winline -Wwrite-strings -Wno-unused-parameter -Wfloat-equal -fsanitize=undefined -fsanitize=unreachable -fsanitize=leak -fsanitize=enum -fsanitize=builtin -fsanitize=shift-exponent -fsanitize=pointer-compare -fsanitize=pointer-subtract -fsanitize=address -Werror
 
 
-OBJECTS = obj/main.o obj/construction.o obj/geoapp.o obj/geo.o obj/input.o obj/geoImpl.o obj/inputSFML.o 
+OBJECTS = obj/main.o obj/construction.o obj/geoapp.o obj/geo.o obj/input.o obj/inputSFML.o 
 bin/app: $(OBJECTS)
 	g++ $(OBJECTS) -o bin/app -lsfml-window -lsfml-system -lsfml-graphics $(CFLAGS)
 
@@ -20,11 +20,8 @@ obj/inputSFML.o: source/inputSFML.cpp source/inputSFML.h source/Input.h
 obj/geo.o: source/geo.cpp source/geo.h source/vectorHolder.h
 	g++ source/geo.cpp -o obj/geo.o -c $(CFLAGS)
 
-obj/geoImpl.o: source/geoImpl.cpp source/geoImpl.h source/geo.h source/vectorHolder.h
-	g++ source/geoImpl.cpp -o obj/geoImpl.o -c $(CFLAGS)
-
 obj/input.o: source/Input.cpp source/Input.h
 	g++ source/Input.cpp -o obj/input.o -c $(CFLAGS)
 
 clean:
-	rm $(OBJECTS)
+	rm obj/*.o
