+++ source/geo.h
@@ -7,27 +7,18 @@
 
 
 //lista figur geometrycznych
+class Shape;
+
 class SegmentShape;
 class TriangleShape;
 class LineShape;
 class CircleShape;
 class PointShape;
-class Shape;
 
 inline double doubleAbs (double r) {
     return r >= 0 ? r : -r;
 }
 
-/*
-struct constructionElements {
-    std::vector<PointShape*> points;
-    std::vector<LineShape*> lines;
-    std::vector<CircleShape*> circles;
-    std::vector<SegmentShape*> segments;
-};
-
-void resetConstructionElements (constructionElements& el);
-*/
 
 typedef vectorHolder <PointShape*, LineShape*, CircleShape*, SegmentShape*, TriangleShape*> constructionElements;
 
@@ -79,17 +70,52 @@ double operator*(const Point &p1, const Point &p2);
 double operator%(const Point &p1, const Point &p2);
 bool operator==(const Point &p1, const Point &p2);
 
-class PointShape : public Shape {
+class PointShape : public Shape{
+    private:
+        static constexpr double radiusOfDrawing=3;
+        Point coordinates;
+        bool exists = true;
+        bool isActive = false;
+        bool isCurrent = false;
+        bool isDependent = false;
     public:
         virtual ~PointShape () {}
 
-        virtual double getX () const = 0;
-        virtual double getY () const = 0;
+        virtual void setExistance (bool ex) override;
+        virtual bool getExistance () const override;
+        virtual void setActivity (bool ac) override;
+        virtual bool getActivity () const override;
+        virtual void setCurrent (bool cu) override;
+        virtual bool getCurrent () const override;
+        virtual void setDependent (bool de) override;
+        virtual bool getDependent () const override;
+        virtual double getX () const;
+        virtual double getY () const;
 
-        virtual void setX (double newX) = 0;
-        virtual void setY (double newY) = 0;
+        virtual void setX (double newX);
+        virtual void setY (double newY);
 
-        virtual double abs() const = 0;
+        virtual void draw(sf::RenderWindow*, const sf::FloatRect& visible, const sf::FloatRect& box) const override;
+
+        virtual void hull_draw(sf::RenderWindow*, const sf::FloatRect& visible, const sf::FloatRect& box) const override;
+
+        virtual double distFromPoint(const Point&) const override;
+
+        PointShape(double=0, double=0);
+
+        virtual double abs() const;
+
+        virtual void addToConstructionElements (constructionElements& el) override;
+        virtual void removeFromConstructionElements (constructionElements& el) override;
+
+        virtual void addToCurrentConditions (uiOptionConditions& op, int c) override;
+        virtual bool isHit (const Point& p) override;
+
+        virtual unsigned int getHitPriority () override;
+
+        virtual void moveShape (double xMov, double yMov) override;
+
+        PointShape (const LineShape&,const LineShape&);
 };
 
 PointShape* makePointShape (double = 0, double = 0);
@@ -97,22 +123,49 @@ PointShape* makePointShape (const LineShape&, const LineShape&);
 
 Point getPointLocation (PointShape&);
 
-class SegmentShape : public Shape{
+class SegmentShape : public Shape {
+    private:
+        Point p1, p2;
+        bool exists = true;
+        bool isActive = false;
+        bool isCurrent = false;
+        bool isDependent = false;
     public:
         virtual ~SegmentShape () {}
 
-        virtual double getFromX () const = 0;
-        virtual double getToX () const = 0;
-        virtual double getFromY () const = 0;
-        virtual double getToY () const = 0;
+        virtual void setExistance (bool ex) override;
+        virtual bool getExistance () const override;
+        virtual void setActivity (bool ac) override;
+        virtual bool getActivity () const override;
+        virtual void setCurrent (bool cu) override;
+        virtual bool getCurrent () const override;
+        virtual void setDependent (bool de) override;
+        virtual bool getDependent () const override;
 
-        virtual void setFromX (double newX) = 0;
-        virtual void setToX (double newX) = 0;
-        virtual void setFromY (double newY) = 0;
-        virtual void setToY (double newY) = 0;
+        virtual double getFromX () const;
+        virtual double getToX () const;
+        virtual double getFromY () const;
+        virtual double getToY () const;
 
+        virtual void setFromX (double newX);
+        virtual void setToX (double newX);
+        virtual void setFromY (double newY);
+        virtual void setToY (double newY);
 
-        virtual double abs() const = 0;
+        SegmentShape(const Point&, const Point&);
+        SegmentShape () {}
+
+        virtual double distFromPoint(const Point&) const override;
+
+        virtual void draw(sf::RenderWindow*, const sf::FloatRect& visible, const sf::FloatRect& box) const override;
+
+        virtual double abs() const;
+        virtual void addToConstructionElements (constructionElements& el) override;
+        virtual void removeFromConstructionElements (constructionElements& el) override;
+        virtual void addToCurrentConditions (uiOptionConditions& op, int c) override;
+        virtual bool isHit (const Point& p) override;
+        virtual unsigned int getHitPriority () override;
+
 };
 
 SegmentShape* makeSegmentShape ();
@@ -121,19 +174,51 @@ SegmentShape* makeSegmentShape (const Point&, const Point&);
 Point getSegmentFrom (const SegmentShape& seg);
 Point getSegmentTo (const SegmentShape& seg);
 
-class LineShape : public Shape{
+class LineShape : public Shape {
+    private:
+        Point n;
+        double c;
+        bool exists = true;
+        bool isActive = false;
+        bool isCurrent = false;
+        bool isDependent = false;
+        void goThroughPoints (const Point& p, const Point& q);
     public:
         virtual ~LineShape () {}
-        virtual double getNormalX () const = 0;
-        virtual double getNormalY () const = 0;
-        virtual double getC () const = 0;
+        virtual void setExistance (bool ex) override;
+        virtual bool getExistance () const override;
+        virtual void setActivity (bool ac) override;
+        virtual bool getActivity () const override;
+        virtual void setCurrent (bool cu) override;
+        virtual bool getCurrent () const override;
+        virtual void setDependent (bool de) override;
+        virtual bool getDependent () const override;
 
-        virtual void setNormalX (double x) = 0;
-        virtual void setNormalY (double y) = 0;
-        virtual void setC (double _c) = 0;
+        virtual double getNormalX () const;
+        virtual double getNormalY () const;
+        virtual double getC () const;
 
-        virtual void goThroughPoints (const double px, const double py, const double qx, const double qy) = 0;
+        virtual void setNormalX (double x);
+        virtual void setNormalY (double y);
+        virtual void setC (double _c);
 
+        LineShape(double,double,double); //line ax+by+c=0
+        LineShape(const Point&,const Point&); //line through two points
+        LineShape(const SegmentShape&);
+
+        double distFromPoint(const Point&) const override;
+        virtual void draw(sf::RenderWindow*, const sf::FloatRect& visible, const sf::FloatRect& box) const override;
+
+
+        virtual void goThroughPoints (const double px, const double py, const double qx, const double qy);
+
+        LineShape(const CircleShape&,const CircleShape&);
+        virtual void addToConstructionElements (constructionElements& el) override;
+        virtual void removeFromConstructionElements (constructionElements& el) override;
+        virtual void addToCurrentConditions (uiOptionConditions& op, int c) override;
+        virtual bool isHit (const Point& p) override;
+
+        virtual unsigned int getHitPriority () override;
 };
 
 LineShape* makeLineShape (double, double, double);
@@ -142,17 +227,47 @@ LineShape* makeLineShape (const Point&, const Point&);
 
 Point getLineNormal (const LineShape&);
 
+
 class CircleShape : public Shape {
+        Point middle;
+        double r;
+        bool exists = true;
+        bool isActive = false;
+        bool isCurrent = false;
+        bool isDependent = false;
     public:
+
         virtual ~CircleShape () {}
-        virtual double getMiddleX () const = 0;
-        virtual double getMiddleY () const = 0;
-        virtual double getR () const = 0;
+        
+        virtual void setExistance (bool ex) override;
+        virtual bool getExistance () const override;
+        virtual void setActivity (bool ac) override;
+        virtual bool getActivity () const override;
+        virtual void setCurrent (bool cu) override;
+        virtual bool getCurrent () const override;
+        virtual void setDependent (bool de) override;
+        virtual bool getDependent () const override;
 
-        virtual void setMiddleX (double x) = 0;
-        virtual void setMiddleY (double y) = 0;
-        virtual void setR (double _r) = 0;
+        virtual double getMiddleX () const;
+        virtual double getMiddleY () const;
+        virtual double getR () const;
 
+        virtual void setMiddleX (double x);
+        virtual void setMiddleY (double y);
+        virtual void setR (double _r);
+
+        virtual double distFromPoint(const Point&) const override;
+        virtual void draw(sf::RenderWindow*, const sf::FloatRect& visible, const sf::FloatRect& box) const override;
+        virtual void hull_draw(sf::RenderWindow*, const sf::FloatRect& visible, const sf::FloatRect& box) const override;
+        CircleShape(const Point&, const Point&, const Point&);
+        CircleShape(const Point&, double);
+        CircleShape(const Point&, const Point&);
+        CircleShape (double, double, double);
+        virtual void addToConstructionElements (constructionElements& el) override;
+        virtual void removeFromConstructionElements (constructionElements& el) override;
+        virtual void addToCurrentConditions (uiOptionConditions& op, int c) override;
+        virtual bool isHit (const Point& p) override;
+        virtual unsigned int getHitPriority () override;
 };
 
 CircleShape* makeCircleShape (const Point&, double);
@@ -163,27 +278,53 @@ CircleShape* makeCircleShape (const Point&, const Point&, const Point&);
 Point getCircleCenter (CircleShape&);
 
 //TODO: Triangle class
+
 class TriangleShape : public Shape {
+        Point A, B, C;
+        bool exists = true;
+        bool isActive = false;
+        bool isCurrent = false;
+        bool isDependent = false;
     public:
+
         virtual ~TriangleShape () {}
-        virtual double getAX () const = 0;
-        virtual double getAY () const = 0;
-
-        virtual double getBX () const = 0;
-        virtual double getBY () const = 0;
         
-        virtual double getCX () const = 0;
-        virtual double getCY () const = 0;
+        virtual void setExistance (bool ex) override;
+        virtual bool getExistance () const override;
+        virtual void setActivity (bool ac) override;
+        virtual bool getActivity () const override;
+        virtual void setCurrent (bool cu) override;
+        virtual bool getCurrent () const override;
+        virtual void setDependent (bool de) override;
+        virtual bool getDependent () const override;
 
-        virtual void setAX (double) = 0;
-        virtual void setAY (double) = 0;
+        virtual double getAX () const;
+        virtual double getAY () const;
+        virtual double getBX () const;
+        virtual double getBY () const;
+        virtual double getCX () const;
+        virtual double getCY () const;
 
-        virtual void setBX (double) = 0;
-        virtual void setBY (double) = 0;
+        virtual void setAX (double _x);
+        virtual void setAY (double _y);
+        virtual void setBX (double _x);
+        virtual void setBY (double _y);
+        virtual void setCX (double _x);
+        virtual void setCY (double _y);
         
-        virtual void setCX (double) = 0;
-        virtual void setCY (double) = 0;
-        
+
+        virtual double distFromPoint(const Point&) const override;
+        virtual void draw(sf::RenderWindow*, const sf::FloatRect& visible, const sf::FloatRect& box) const override;
+        virtual void hull_draw(sf::RenderWindow*, const sf::FloatRect& visible, const sf::FloatRect& box) const override;
+        TriangleShape(const Point&, const Point&, const Point&);
+        TriangleShape (double, double, double, double, double, double);
+        virtual void addToConstructionElements (constructionElements& el) override;
+        virtual void removeFromConstructionElements (constructionElements& el) override;
+        virtual void addToCurrentConditions (uiOptionConditions& op, int c) override;
+        virtual bool isHit (const Point& p) override;
+        virtual unsigned int getHitPriority () override;
 };
 TriangleShape* makeTriangleShape(const Point&, const Point&, const Point&);
 TriangleShape* makeTriangleShape(double, double, double, double, double, double);
+
+sf::Color getShapeColor (bool active, bool current, bool dependent);
